feature_patch_4
Added
* Functions can now return values... Um.. Well not exactly return values but instead they inject these variables into the caller!
* Now THATLANG only supports int, double, string and boolean types.
* These operators are too tiring to make.. The code base currently contains two deprecated experiments, and one prototype. I hopw some day i'll find a new and easy to expand operators system

+++ src/main/java/in/mcxiv/interp/functions/InputFunctions.java
@@ -39,7 +39,7 @@ public class InputFunctions extends FunctionEvaluator {
         String s = Try.If(scanner::hasNextLine).Then(scanner::nextLine).Else(() -> "").toString();
         THATObject variable = THOSEObjects.createValue(switch (fct.getValue()) {
             case SCAN_LINE -> s;
-            case SCAN_INT -> PrimitiveParser.LONG.parse(s);
+            case SCAN_INT -> PrimitiveParser.INT.parse(s);
             case SCAN_FLOAT -> PrimitiveParser.DOUBLE.parse(s);
             default -> throw new IllegalStateException("Unexpected value: " + fct.getValue());
         });
